requirements.txt
pre-commit==3.7.1
ruff==0.4.4
pyyaml==6.0.1
psutil==5.9.8
watchdog==3.0.0
jsonschema==4.22.0
